+++ src/stores/auth.js
@@ -1,5 +1,5 @@
 import { defineStore } from 'pinia';
-import axios from '../axios';
+import axios from '../core/axios';
 
 export default defineStore('auth', {
     state: () => {
